refactor(admin-cb): move block preview const to separate file
