+++ src/admin/content-block/components/ContentBlockPreview/ContentBlockPreview.tsx
@@ -2,21 +2,7 @@ import classnames from 'classnames';
 import React, { FunctionComponent } from 'react';
 import { RouteComponentProps, withRouter } from 'react-router-dom';
 
-import {
-	BlockAccordions,
-	BlockButtons,
-	BlockCTAs,
-	BlockGrid,
-	BlockHeading,
-	BlockIFrame,
-	BlockImage,
-	BlockIntro,
-	BlockProjectsSpotlight,
-	BlockRichText,
-	ButtonAction,
-	Container,
-	Spacer,
-} from '@viaa/avo2-components';
+import { ButtonAction, Container, Spacer } from '@viaa/avo2-components';
 import { Avo } from '@viaa/avo2-types';
 
 import { navigateToContentType } from '../../../../shared/helpers';
@@ -31,11 +17,10 @@ import {
 	DARK_BACKGROUND_COLOR_OPTIONS,
 } from '../../content-block.const';
 import {
-	MediaGridWrapper,
-	MediaPlayerTitleTextButtonWrapper,
-	MediaPlayerWrapper,
-	PageOverviewWrapper,
-} from '../wrappers';
+	COMPONENT_PREVIEW_MAP,
+	IGNORE_BLOCK_LEVEL_PROPS,
+	REPEATABLE_CONTENT_BLOCKS,
+} from './ContentBlockPreview.const';
 
 interface ContentBlockPreviewProps extends RouteComponentProps {
 	componentState: ContentBlockComponentState | ContentBlockComponentState[];
@@ -48,40 +33,6 @@ enum ContentWidthMap {
 	LARGE = 'large',
 	MEDIUM = 'medium',
 }
-
-const COMPONENT_PREVIEW_MAP = Object.freeze({
-	[ContentBlockType.Accordions]: BlockAccordions,
-	[ContentBlockType.CTAs]: BlockCTAs,
-	[ContentBlockType.Buttons]: BlockButtons,
-	[ContentBlockType.Heading]: BlockHeading,
-	[ContentBlockType.IFrame]: BlockIFrame,
-	[ContentBlockType.Intro]: BlockIntro,
-	[ContentBlockType.Image]: BlockImage,
-	[ContentBlockType.MediaGrid]: MediaGridWrapper,
-	[ContentBlockType.MediaPlayer]: MediaPlayerWrapper,
-	[ContentBlockType.MediaPlayerTitleTextButton]: MediaPlayerTitleTextButtonWrapper,
-	[ContentBlockType.RichText]: BlockRichText,
-	[ContentBlockType.RichTextTwoColumns]: BlockRichText,
-	[ContentBlockType.IFrame]: BlockIFrame,
-	[ContentBlockType.Accordions]: BlockAccordions,
-	[ContentBlockType.Image]: BlockImage,
-	[ContentBlockType.ImageGrid]: BlockGrid,
-	[ContentBlockType.PageOverview]: PageOverviewWrapper,
-	[ContentBlockType.ProjectsSpotlight]: BlockProjectsSpotlight,
-});
-
-export const REPEATABLE_CONTENT_BLOCKS = [
-	ContentBlockType.Accordions,
-	ContentBlockType.Buttons,
-	ContentBlockType.CTAs,
-	ContentBlockType.RichText,
-	ContentBlockType.RichTextTwoColumns,
-	ContentBlockType.MediaGrid,
-	ContentBlockType.ImageGrid,
-	ContentBlockType.ProjectsSpotlight,
-];
-
-const IGNORE_BLOCK_LEVEL_PROPS = ['position', 'elements', 'blockType'];
 
 const ContentBlockPreview: FunctionComponent<ContentBlockPreviewProps> = ({
 	history,
